% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_melt.R
\name{fast_melt}
\alias{fast_melt}
\title{A helper function}
\usage{
fast_melt(physeq)
}
\arguments{
\item{physeq}{(Required). A \code{phyloseq} object containing abundance table,
taxonomic assignment, sample data and/or phylogenetic tree if available.}
}
\value{
A long \code{data.frame} object.
}
\description{
This function takes a \code{phyloseq} object and melts the otu table
into a OTU_name, Prevalence, TotalCount long format table.
}
\examples{
data(GlobalPatterns)
GP <- fast_melt(GlobalPatterns)
}
