% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSprep.R
\name{FSprep}
\alias{FSprep}
\title{A function to prepare data for FastSpar}
\usage{
FSprep(physeq, condition, prev = 0.1)
}
\arguments{
\item{physeq}{(Required). A \code{phyloseq} object containing abundance table,
taxonomic assignment, sample data and/or phylogenetic tree if available.}

\item{condition}{(Required). A condition that is compatible with \code{phyloseq} object.
A metadata header == metadata state. E.g. \code{Health==healthy}}

\item{prev.}{Prevalence of an OTU as a fraction of the total number of samples. Default = 0.1}
}
\value{
A \code{data.frame} object with \verb{#OTU ID} inserted in the first cell.
}
\description{
Creates a data.frame from the original \code{phyloseq} object
with the relevant conditions satisfied.
}
\examples{
data(GlobalPatterns)
sample_data(GlobalPatterns)
GPfeces <- FSprep(GlobalPatterns, SampleType==feces,prev=0.1)
write.table(GPfeces, sep ="\t", file = "Can.txt", quote = FALSE,row.names = FALSE)
}
\author{
Abhishake Lahiri \email{abhishake.lahiri@gmail.com}
}
